hw_2/Code_op1/pyproject.toml
[project]
name = "code-op1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gradio>=6.11.0",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "pywavelets>=1.9.0",
    "scikit-image>=0.26.0",
]
